requirements.txt
# Core AI Frameworks
crewai>=0.1.0
pydantic-ai>=0.1.0
pydantic>=2.0.0

# FastAPI Server (latest)
fastapi[standard]>=0.128.0
uvicorn[standard]>=0.32.0

# STORM - Stanford AI Research & Article Generation
knowledge-storm>=1.1.0

# LLM Provider - Simplified with OpenRouter
openai>=1.0.0  # Single package for ALL models via OpenRouter (Groq, GPT-4, Claude, etc.)
litellm>=1.0.0  # Required by STORM, also supports OpenRouter

# LangChain integrations (required by agents/seo/topical_mesh_architect.py)
langchain-groq>=0.1.0  # Required for ChatGroq in CrewAI agents

# OLD (commented out - uncomment if rollback needed):
# groq>=0.4.0  # Not needed - Groq available via OpenRouter!
# langchain-openai>=0.0.5  # Not needed - using openai SDK directly

# Data Collection & Analysis
exa-py>=1.0.0
firecrawl-py>=0.0.1
beautifulsoup4>=4.12.0
requests>=2.31.0

# SEO & Research Tools
advertools>=0.14.0  # SEO keyword generation, crawling, sitemap analysis
serpapi>=0.1.0
spacy>=3.7.0
networkx>=3.0.0
matplotlib>=3.8.0
python-louvain>=0.16  # Community detection for topical mesh analysis

# Email & Templates
jinja2>=3.1.0
sendgrid>=6.11.0

# Utilities
pytest
pytest-asyncio
pytest-cov

python-dotenv>=1.0.0
pyyaml>=6.0.0
pandas>=2.0.0  # Required by advertools
libsql-client>=0.3.1
